Factor out tree logic
