+++ hypothesis-python/RELEASE.rst
@@ -1,0 +1,4 @@
+RELEASE_TYPE: patch
+
+This is a pure refactoring release that extracts some logic from the core Hypothesis engine
+into its own class and file. It should have no user visible impact.
